Fix sha256sum is not available on Mac OS X (#70)
Problem:
When running this script on Mac OS X, the script throws an error stating that the SHA does not match.

Cause:
The SHA is not actually being compared. The problem is that the sha256sum program is not available on Mac OS X, so it errors out.

Fix:
Use `shasum -a 256` instead.