Remove duplicate install_composer script (#73)
* Remove duplicate install_composer script

* Fix bin/hub does not mount host Hub config

* Fix paths in install_composer

Problem:
The path to the composer file is hard-coded. This means that if the working directory is not the root of the project when this script is run, then composer will be installed in the wrong place.

Fix:
Use the existing ROOT env var to ensure composer is installed in the project's bin directory.